pyproject.toml
[tool.poetry]
name = "tutor-gpt"
version = "0.1.0"
description = "LangChain LLM application. Dynamic few-shot metaprompting for the task of tutoring."
authors = ["vintro <vince@plasticlabs.ai>", "VVoruganti <vineeth@plasticlabs.ai"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "0.0.251"
py-cord = "^2.4.1"
python-dotenv = "^1.0.0"
validators = "^0.20.0"
openai = "^0.27.5"
pexpect = "^4.8.0"
tiktoken = "^0.3.3"
wolframalpha = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
